refactor : user data domain 레이어 리팩토링
