+++ lib/features/user/usecases/resign_user_info_use_case.dart
@@ -17,6 +17,6 @@ final class ResignUserInfoUseCase
     final topicRes = await _topicRepository.deleteUserWrongAnswers();
     topicRes.getOrThrow();
 
-    return _userRepository.quitUser(request);
+    return _userRepository.deleteUser(request);
   }
 }
